# Fix regexp set matches for literal matchers