While testing a feature for Adaptive Metrics, I was surprised to find out that posting lookups for `foo=~"(bar|bar)"` are faster than `foo=~"bar"`. It turns out I introduced a performance regression in https://github.com/grafana/mimir-prometheus/pull/463.

When we added the `optimizeAlternatingLiterals` function, we subtly broke one edge case. A regexp matcher which matches a single literal, like `foo=~"bar"` used to return `bar` from `SetMatches()`, but currently does not. The implication is that the tsdb will first do a LabelValues call to get all values for `foo`, then match them against the regexp `bar`. This PR restores the previous behavior which is able to directly lookup postings for `foo="bar"` instead. In a random sample of 1M queries from a large dedicated cell, about 11% of queries had at least one matcher in this form.

I've added a new benchmark to show the difference for a label with high cardinality:

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb
                                                    │   before.txt    │              after.txt              │
                                                    │     sec/op      │   sec/op     vs base                │
Querier/Head/PostingsForMatchers/i_times_n=~"0"-10    7136391.5n ± 4%   362.0n ± 1%  -99.99% (p=0.000 n=10)
Querier/Block/PostingsForMatchers/i_times_n=~"0"-10     7578.47µ ± 0%   49.29µ ± 1%  -99.35% (p=0.000 n=10)
geomean                                                   7.354m        4.224µ       -99.94%

                                                    │   before.txt    │              after.txt              │
                                                    │      B/op       │    B/op     vs base                 │
Querier/Head/PostingsForMatchers/i_times_n=~"0"-10    7372864.00 ± 0%   64.00 ± 0%  -100.00% (p=0.000 n=10)
Querier/Block/PostingsForMatchers/i_times_n=~"0"-10   7372896.00 ± 0%   80.00 ± 0%  -100.00% (p=0.000 n=10)
geomean                                                  7.031Mi        71.55       -100.00%

                                                    │ before.txt │             after.txt              │
                                                    │ allocs/op  │ allocs/op   vs base                │
Querier/Head/PostingsForMatchers/i_times_n=~"0"-10    4.000 ± 0%   3.000 ± 0%  -25.00% (p=0.000 n=10)
Querier/Block/PostingsForMatchers/i_times_n=~"0"-10   6.000 ± 0%   4.000 ± 0%  -33.33% (p=0.000 n=10)
geomean                                               4.899        3.464       -29.29%

```